+++ videocall-client/src/connection/webmedia.rs
@@ -44,7 +44,6 @@ pub enum TransportHint {
 
 #[derive(Clone)]
 pub struct ConnectOptions {
-    pub userid: String,
     pub websocket_url: String,
     pub webtransport_url: String,
     pub on_inbound_media: Callback<PacketWrapper>,