+++ bot/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.3](https://github.com/security-union/videocall-rs/compare/bot-v1.1.2...bot-v1.1.3) - 2026-02-23
+
+### Other
+
+- updated the following local packages: videocall-types
+
 ## [1.1.2](https://github.com/security-union/videocall-rs/compare/bot-v1.1.1...bot-v1.1.2) - 2026-02-19
 
 ### Other